seperate asm and asm's main
